test(store): move the migration runner tests in-package
